# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API server for GoTicket ticketing system.",
        "title": "GoTicket API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "GoTicket Support Team",
            "url": "http://www.goticket.com/support",
            "email": "support@goticket.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:18080",
    "basePath": "/api/v1",
    "paths": {
...
        "/concerts": {
            "get": {
                "description": "Retrieve a paginated list of concerts with optional search and sorting",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Concerts"
                ],
                "summary": "List concerts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page limit (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword for title, venue, or description",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting order (title_asc, title_desc, date_asc, date_desc)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.ConcertResponse"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/dto.PaginationMeta"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.WebResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create concert data along with file uploads for poster, thumbnail, and rules PDF",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Concerts"
                ],
                "summary": "Create a new concert",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of concert",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Venue of concert",
                        "name": "venue",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date of concert (RFC3339 or YYYY-MM-DD HH:mm)",
                        "name": "date",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Description of concert",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Status of concert",
                        "name": "status",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Poster Image file",
                        "name": "poster",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Thumbnail Image file",
                        "name": "thumbnail",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Concert Rules PDF file",
                        "name": "rules_pdf",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ConcertResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.WebResponse"
                        }
                    }
                }
            }
        },
        "/concerts/{id}": {
            "get": {
                "description": "Retrieve single concert details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Concerts"
                ],
                "summary": "Get concert by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Concert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ConcertResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.WebResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.WebResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update existing concert data and optionally upload new poster, thumbnail, or rules PDF",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Concerts"
                ],
                "summary": "Update concert by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Concert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title of concert",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Venue of concert",
                        "name": "venue",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Date of concert",
                        "name": "date",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Description of concert",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Status of concert",
                        "name": "status",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Poster Image file",
                        "name": "poster",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Thumbnail Image file",
                        "name": "thumbnail",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Concert Rules PDF file",
                        "name": "rules_pdf",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ConcertResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.WebResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a concert record from database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Concerts"
                ],
                "summary": "Delete concert by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Concert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.WebResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ConcertResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "poster_url": {
                    "type": "string"
                },
                "rules_pdf_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "venue": {
                    "type": "string"
                }
            }
        },
...
        "dto.PaginationMeta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "dto.WebResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer ' followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}